+++ lib/storage/local/progress_store.g.dart
@@ -10,34 +10,79 @@ part of 'progress_store.dart';
 // ignore_for_file: type=lint, type=warning
 
 @ProviderFor(ProgressStore)
-final progressStoreProvider = ProgressStoreProvider._();
+final progressStoreProvider = ProgressStoreFamily._();
 
 final class ProgressStoreProvider
     extends $StreamNotifierProvider<ProgressStore, Map<String, MediaProgress>> {
-  ProgressStoreProvider._()
-    : super(
-        from: null,
-        argument: null,
-        retry: null,
-        name: r'progressStoreProvider',
-        isAutoDispose: false,
-        dependencies: null,
-        $allTransitiveDependencies: null,
-      );
+  ProgressStoreProvider._({
+    required ProgressStoreFamily super.from,
+    required String super.argument,
+  }) : super(
+         retry: null,
+         name: r'progressStoreProvider',
+         isAutoDispose: false,
+         dependencies: null,
+         $allTransitiveDependencies: null,
+       );
 
   @override
   String debugGetCreateSourceHash() => _$progressStoreHash();
 
+  @override
+  String toString() {
+    return r'progressStoreProvider'
+        ''
+        '($argument)';
+  }
+
   @$internal
   @override
   ProgressStore create() => ProgressStore();
+
+  @override
+  bool operator ==(Object other) {
+    return other is ProgressStoreProvider && other.argument == argument;
+  }
+
+  @override
+  int get hashCode {
+    return argument.hashCode;
+  }
 }
 
-String _$progressStoreHash() => r'cac76f991844e9eec3ff13b75987e68417883b84';
+String _$progressStoreHash() => r'9a1a968d6636d15b2e2d029571b1da807eb2217d';
+
+final class ProgressStoreFamily extends $Family
+    with
+        $ClassFamilyOverride<
+          ProgressStore,
+          AsyncValue<Map<String, MediaProgress>>,
+          Map<String, MediaProgress>,
+          Stream<Map<String, MediaProgress>>,
+          String
+        > {
+  ProgressStoreFamily._()
+    : super(
+        retry: null,
+        name: r'progressStoreProvider',
+        dependencies: null,
+        $allTransitiveDependencies: null,
+        isAutoDispose: false,
+      );
+
+  ProgressStoreProvider call(String userId) =>
+      ProgressStoreProvider._(argument: userId, from: this);
+
+  @override
+  String toString() => r'progressStoreProvider';
+}
 
 abstract class _$ProgressStore
     extends $StreamNotifier<Map<String, MediaProgress>> {
-  Stream<Map<String, MediaProgress>> build();
+  late final _$args = ref.$arg as String;
+  String get userId => _$args;
+
+  Stream<Map<String, MediaProgress>> build(String userId);
   @$mustCallSuper
   @override
   WhenComplete runBuild() {
@@ -58,6 +103,6 @@ abstract class _$ProgressStore
               Object?,
               Object?
             >;
-    return element.handleCreate(ref, build);
+    return element.handleCreate(ref, () => build(_$args));
   }
 }