+++ CHANGELOG.md
@@ -7,6 +7,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Fixed
+
+- unsupported device language fallback properly to english
+
 ## [v0.6.2] - 2026-08-23
 
 ### Fixed