+++ CHANGELOG.md
@@ -7,6 +7,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+## [v0.6.3] - 2026-08-23
+
 ### Added
 
 - Polish translations
@@ -565,7 +567,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Custom fonts support
 - Dynamic theme support
 
-[Unreleased]: https://github.com/likhithpraveenk/storii/compare/v0.6.2...HEAD
+[Unreleased]: https://github.com/likhithpraveenk/storii/compare/v0.6.3...HEAD
+[v0.6.3]: https://github.com/likhithpraveenk/storii/compare/v0.6.2...v0.6.3
 [v0.6.2]: https://github.com/likhithpraveenk/storii/compare/v0.6.1...v0.6.2
 [v0.6.1]: https://github.com/likhithpraveenk/storii/compare/v0.6.0...v0.6.1
 [v0.6.0]: https://github.com/likhithpraveenk/storii/compare/v0.5.14...v0.6.0